requirements.txt
boto3==1.26.62
psycopg2==2.9.5
python-dotenv==0.21.1
